feat: login google and cart

+++ database/migrations/2024_09_17_131706_create_orders_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->decimal('total_price', 10, 2);
+            $table->integer('total_price');
             $table->enum('status', ['pending', 'paid', 'shipped', 'delivered', 'canceled']);
             $table->timestamps();
         });
